# ENH: Improve performance of ndimage.binary_erosion