For smaller input of the `binary_erosion` method we improve performance by simplifying a check on the size of the array. We also avoid creation of a temporary array in `_center_is_true`

Benchmark:
```
import numpy as np
import scipy.ndimage as ndimage

data  = (np.random.rand(24, 20)+.5).astype(int)
structure=np.array([[1,1,1]])
%timeit ndimage.binary_erosion(data, structure)
```
Results:
```
main: 8.45 µs ± 273 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
PR random data 5.47 µs ± 48.3 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
```

<!-- 
Thanks for contributing a pull request! Please ensure that
your PR satisfies the checklist before submitting:
http://scipy.github.io/devdocs/dev/contributor/development_workflow.html#checklist-before-submitting-a-pr

Also, please name and describe your PR as you would write a
commit message:
http://scipy.github.io/devdocs/dev/contributor/development_workflow.html#writing-the-commit-message

Depending on your changes, you can skip CI operations and save time and energy: 
http://scipy.github.io/devdocs/dev/contributor/continuous_integration.html#skipping

Note that we are a team of volunteers; we appreciate your
patience during the review process.

Again, thanks for contributing!

#### Reference issue

#### What does this implement/fix?

-->

#### Additional information

There are more places in the scipy codebase where we can improve performance by replacing `np.prod` with faster code. In this first PR we only update `scipy.ndimage._morphology`.